refactor: extract rename into features layer + dead code cleanup
- Move rename logic from backend/rename.rs → features/rename.rs
  (backend becomes 30-line thin adapter)
- Remove dead RawCursor struct (features/cursor.rs)
- Remove dead kw_for_kind function (backend/format.rs)
- Remove dead rg_scope_for_file wrapper (backend/mod.rs)
- Remove unused SymbolKind import

All 11 feature modules now extracted. Backend is thin adapter only.
